Don't translate these strings
Maybe someday. The niche of users who would use this plugin is small enough; I'm sorry but the subset of users who would appreciate this plugin's title appearing in their computer's language is so small it probably doesn't even exist. And this code relies on this weird _ function from libpurple/internal that the purple plugin pack is allowed to use because it's so special, but isn't exposed in the libpurple-dev release. So it's just not worth it